libpriv/core: also checksum package action
When calculating the goal checksum, we only accounted for package
installs. But with override support, there are now other package actions
we need to pick up on.

In practice, the fact that we also checksum the treespec made this hard
to actually cause an issue. Although, since the actual final goal is
also dependent on the underlying rpmdb, I can imagine e.g. a replacement
override be considered a `DNF_PACKAGE_INFO_UPDATE` for one rpmdb appear
as a `DNF_PACKAGE_INFO_DOWNGRADE` for another.

While we're here, do some light code style porting and streamline the
checksumming process to avoid creating a separate `GPtrArray`.

Closes: #1053
Approved by: cgwalters