Revert "tests/libvm: write rpm build logs to file"
This reverts commit 1ef259ed76119d50d7e82e3cc013263f3e662625.

Basically, for historical POSIX compatibility, `errexit` doesn't work at
all inside functions called from an if-statement/boolean context. This
is something I had already learned (and forgotten) when hacking on PAPR.

There are ways around this, but it's just not worth the added complex
shell goop for what it gives us.

http://mywiki.wooledge.org/BashFAQ/105